agent_stdio/requirements.txt
google-adk
fastmcp
mcp
requests
python-dotenv
